render relay data on server using sync-request, maybe; deterministic seeds
