+++ app/javascript/server-polyfills/storage.js
@@ -1,16 +1,12 @@
 // @flow
 function storage () {
-  let data: { [key: string]: string } = {}
-
   return {
     getItem (key: string) {
-      return data[key] === undefined ? null : data[key]
+      return null
     },
     setItem (key: string, value: any) {
-      data[key] = value + '' // force to string
     },
     removeItem (key: string) {
-      delete data[key]
     }
   }
 }
