+++ app/javascript/src/shared/relay/createQuery/createQuery.client.js
@@ -1,14 +1,17 @@
 // @flow
 import { QueryResponseCache } from 'relay-runtime'
 import type { QueryResult, QueryPayload, FetchFunction } from 'relay-runtime'
-import { createBaseQuery } from './createBaseQuery'
 import { getCacheResolver } from './cacheResolvers'
+import config from 'shared/config'
 
 // see: https://facebook.github.io/relay/docs/network-layer.html
 // see: https://github.com/facebook/relay/issues/1688#issuecomment-302931855
-export function createQuery (headers: Object): FetchFunction {
+export function createQuery (extraHeaders: Object): FetchFunction {
   const cache = new QueryResponseCache({ size: 250, ttl: 60 * 5 * 1000 })
-  const baseQuery = createBaseQuery(headers)
+  const headers = {
+    'content-type': 'application/json',
+    ...extraHeaders
+  }
 
   return async function query (operation, variables) {
     // find an applicable stop-gap cache resolver (if any)
@@ -29,7 +32,16 @@ export function createQuery (headers: Object): FetchFunction {
     }
 
     // fetch data from network if missed
-    const result = await baseQuery(operation, variables)
+    const response = await fetch(config.graphUrl, {
+      method: 'POST',
+      headers,
+      body: JSON.stringify({
+        query: operation.text,
+        variables
+      })
+    })
+
+    const result = await response.json()
 
     // cache response payload if success
     payload = asPayload(result)
