+++ app/javascript/src/shared/relay/createQuery/createQuery.server.js
@@ -1,11 +1,25 @@
 // @flow
+import request from 'sync-request'
 import type { FetchFunction } from 'relay-runtime'
-import { createBaseQuery } from './createBaseQuery'
+import config from 'shared/config'
 
-export function createQuery (headers: Object): FetchFunction {
-  const baseQuery = createBaseQuery(headers)
+export function createQuery (extraHeaders: Object): FetchFunction {
+  const headers = {
+    'content-type': 'application/json',
+    ...extraHeaders
+  }
 
   return function query (operation, variables) {
-    return baseQuery(operation, variables)
+    const response = request('POST', config.graphUrl, {
+      headers,
+      json: {
+        query: operation.text,
+        variables
+      }
+    })
+
+    const result = JSON.parse(response.getBody('utf8'))
+
+    return result
   }
 }
