+++ db/seeds.rb
@@ -10,7 +10,17 @@ committees = chambers.flat_map do |chamber|
 end
 
 hearings = committees.flat_map do |committee|
-  create_list(:hearing, 2, committee: committee)
+  hearing1 = create(:hearing,
+    committee: committee,
+    date: Faker::Time.between(Time.zone.today, 1.week.from_now)
+  )
+
+  hearing2 = create(:hearing,
+    committee: committee,
+    date: Faker::Time.between(1.week.from_now, 1.month.from_now)
+  )
+
+  [hearing1, hearing2]
 end
 
 hearings.flat_map do |hearing|
