+++ package.json
@@ -39,11 +39,12 @@
     "react-copy-to-clipboard": "^5.0.0",
     "react-dom": "^16.1.1",
     "react-fontawesome": "^1.6.1",
-    "react-relay": "^1.1.0",
+    "react-relay": "^1.4.1",
     "react-router-dom": "^4.1.2",
     "react-router-scroll-4": "^1.0.0-beta.1",
     "react-sticky": "6.0.1",
-    "react-transition-group": "1.x"
+    "react-transition-group": "1.x",
+    "sync-request": "^4.1.0"
   },
   "devDependencies": {
     "@dhau/relay-compiler-webpack-plugin": "^0.5.5",
